requirements.txt
ursina==7.0.0
panda3d==1.10.13
pillow>=10.0.0
